requirements.txt
mysql-connector-python==8.0.13
python-telegram-bot
bs4
gunicorn
flask
